tests/test_p0048_rotate_image.py
import pytest
from leetcode.p0048_rotate_image import rotate, rotate2


@pytest.mark.parametrize(
    ("arr", "rotated"),
    (
        (
            [[1, 2, 3, 4], [5, 6, 7, 8], [9, 10, 11, 12], [13, 14, 15, 16]],
            [[13, 9, 5, 1], [14, 10, 6, 2], [15, 11, 7, 3], [16, 12, 8, 4]],
        ),
        ([[1, 2, 3], [4, 5, 6], [7, 8, 9]], [[7, 4, 1], [8, 5, 2], [9, 6, 3]]),
    ),
)
def test_rotate(arr, rotated):
    rotate(arr)
    assert arr == rotated


@pytest.mark.parametrize(
    ("arr", "rotated"),
    (
        (
            [[1, 2, 3, 4], [5, 6, 7, 8], [9, 10, 11, 12], [13, 14, 15, 16]],
            [[13, 9, 5, 1], [14, 10, 6, 2], [15, 11, 7, 3], [16, 12, 8, 4]],
        ),
        ([[1, 2, 3], [4, 5, 6], [7, 8, 9]], [[7, 4, 1], [8, 5, 2], [9, 6, 3]]),
    ),
)
def test_rotate2(arr, rotated):
    rotate2(arr)
    assert arr == rotated
